add TB interface to testhelpers package
Adds TB interface to the testhelpers package and modifies the existing
helpers to use the interface instead of `testing.TB`. This allows for
exporting test helpers from packages without bringing in the `testing`
package and the side effects importing it causes, such as additional
flag definitions in the default flag set.