Regenerated proto files with Go 1.14
Generating the files with of the gitalypb package changed the formatting
of one file, tripping up our CI as it regenerates the files and tests
against an empty diff. By commiting the change into the repository, it's
expected to pass again and we can move forward with the previous commit.