git: stats: Remove accessors for ReferenceDiscovery stats
The ReferenceDiscovery struct has a set of accessors to access the
values of its internal fields. All accessors return the value verbosely,
though, without any processing at all and are thus trivial.

Given that the package is internal anyway, there is no real need for
defensive coding here, so this commit removes the accessors and makes
the variables public instead.