smarthttp: Inject "deepens" Prometheus counter
The "deepens" Prometheus counter is currently managed as a global
variable in the smarthttp server package, which is a code smell. To
improve testability and to get in line with other Prometheus counters,
this commit switches code over to use dependency injection for the
counter instead by extending the `smarthttp.NewServer()` function to
accept options and makes the counter a field of the server.