Increase disk size
The based image default is 10GB which is insufficient for a single
large repository like www-gitlab-com to be repacked, or even moved to a
different shard since the host operating takes up space too.

The gitlab and gitaly-n nodes are now set to 250GB to provide
sufficient space to move a large repo from the default storage after
doing a simple Omnibus install to the HA Gitaly cluster.