golang: Drop support for Golang 1.12
Now both omnibus[1] and CNG[2] use Golang 1.13, we can drop support for
Go 1.12. This would allow us to use 1.13 features if required. Further
it allows us to only support Go versions which are suppported by Google.

This change removes the Go 1.12 jobs from the CI pipeline, and updates
the documentation to reflect that one should use 1.13.9 or up from now
on.

Further, it removes an implementation copied from Go 1.13s reflect
package, and uses the the original now we can depend on it.

[1]: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/5179
[2]: https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/415