balancer: silence deprecation warnings
There are currently two deprecation warnings related to gRPC in the
balancer code for `conn.NewAddress()` and `conn.NewServiceConfig()`
which should both be replaced with calls to `conn.UpdateState()`
instead. The issue with `UpdateState()` though is that you can only
replace the complete state, that is both address and configuration, at
the same time. As we do not have both pieces of information ready when
calling `NewAddress()` and `NewServiceConfig()`, let's silence those
warnings for now to allow the linter to become warning-free.