golangci: Make deprecation warnings a hard failure
We had several deprecation warnings in our code, which is why we
disabled these linting rules in our linter job and instead moved it to a
warnings-only job that is allowed to fail. As all of these warnings have
been fixed by now, let's remove that warnings-only job and re-enable
these linting rules in the main job again, making them a hard
requirement.