Bad test setup as a cause of the race failure
Test setup for checking of replication events processing
was incorrect and it lead to failing tests from time to
time (especially with a race detector).
Use of properly initialized global gitay-ruby instance.
Right order of binaries preparation and config setup.
It is imbossible to use multiple Gitaly instances in one
or multiple tests because of global Config used throw the
project.
gRPC services must be registered before Serve call for it.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2592