Overwrite ssh command for git object existence check
Tests use `cat-file -e` to check if an object exists in the repo.
For partial clones, if object doesn't exist locally git tries to
reach remote. For some tests(TestUploadPackCloneWithPartialCloneFilter)
it was `git@localhost` for CI it most likely, immediately failed
because sshd isn't running on it but on my machines it was asking
for password and hanging tests execution. To workaround this git
ssh command is overwritten to false so as soon as tries to reach
remote it fails.

Signed-off-by: Mateusz Nowotyński <maxmati4@gmail.com>