Update conflict error message in rebase_test.go
In Git 2.26.0, the default rebase backend has been changed from an
'apply' backend to a 'merge' backend. This means that the error
message that appears in case of conflict is different than what it
used to be.

Let's update this error message to the one produced by the 'merge'
backend in `rebase_test.go`.