git: Add a metric counting the number of negotiated packfiles
We have recently added a set of metrics that count which Git features
are used during a git-clone(1) client operation for both the smarthttp
and ssh services. While these metrics can already be put into a relation
to the overall number of clones for smarthttp as the PostUploadPack RPC
is used for packfile generation only, the ssh service does both
reference discovery and packfile negotiation in the SSHUploadPack RPC.
As a result, the number of actual packfile negotiations get drowned out
by reference discoveries, so we can't really tell how many clients use
any of the tracked metrics.

Add a new metric to explicitly track how many packfile negotiations were
performed to be able to tell both kinds of requests apart.