git: Convert packfile negotiation stats to use counters
We currently store the values of "haves", "wants" and "shallow" lines as
part of the packfile negotiation. This may become quite expensive in
case a client decides to send thousands of these lines, which would be
allowed by the Git protocol. As a result, the packfile negotiation stats
would grow with the incoming request. As we currently only ever require
the information about if there were any such lines at all, we can just
convert them to use simple counters.

Note that while we track capabilities in an array as well, they are not
subject to the described issue: only a single line may contain
capabilties and the pkt-line format is bounded to at most 65516 bytes.