output junit report when tests fail
In 5109ed2d, pipefail was turned on in order to produce a correct
exit code on test failure. This caused the JUnit reporter not
to run on test failures. This commit turns off pipefail, ensuring
the JUnit reporter runs even after tests fail and returns the
exit code of 'go test'.