praefect: Implement functions to inject/extract connection details
In order to implement two-phase commits in Gitaly, we need to be able to
establish a communication channel with Praefect in order to notify it
about any new reference transaction that are about to start. To work
towards this goal, this commit introduces the logic to encode connection
information for Praefect into the gRPC metadata and extract it on the
receiving side again.