gitaly-hooks: Extract function to connect to Gitaly
The logic to connect to the Gitaly server from a hook is fairly
self-contained given that the looked-up environment variables are not
being used anywhere else. We can thus neatly extract the logic required
for connecting into its own function to make the code flow more obvious
and easier to follow.