receivepack: Inject Praefect connection details into hooks
In order to implement reference transactions via the pre-receive hook,
we need the hook to know how to connect to Praefect. Let's make this
information available by injecting it into the proxied stream's metadata
and extracting it on the receiving side again when setting up the
receive-pack hook environment variables.

Note: due to the fact that not all Gitaly connections are proxied via
Praefect, we cannot rely on the information being encoded into the
metadata. Any failure to extract the Praefect server connection details
is thus non-fatal.