STYLE: add a note about character sets for coding
Developer-facing code should use the ASCII character set as far as
possible to ensure improve accessibility across platforms. Many
platforms lack a lot of UTF8 code glyps, leading to code that is hard to
read on such platforms if they are used.

Add a note to our style guide that mandates this rule.