ignore repositories deleted during a file walk
When starting a file walk, a directory listing is first taken and then
each of the directories are walked in lexicographical order. If a
directory not yet walked is deleted after the file walk was started,
WalkFunc receives a 'not exists' error. This commit ignores those
errors/directories instead of aborting the walk.