objectpool: Update stale comment for unlinking repos
Previous to commit c02b9ad1 (UnlinkRepositoryFromObjectPool: stop
removing objects/info/alternates, 2019-03-29), we removed the
"objects/info/alternates" file for a repository when unlinking it from
its object pool. While the commit changed it to not delete the file
anymore, it forgot to update a comment that claims we still do.

Update the comment to avoid confusion.