Use separate http settings config object
We want to fall back to the legacy config if we find any empty values in
the http_settings. This is to guard against a partially empty config
value getting passed from gitaly, which has empty fields in the http
settings. This could happen as we transition from using the legacy shell
config yml to the gitaly config toml.