praefect: Extract transaction manager
In order to implement transactions, a new reference transaction service
was created that provides multiple RPCs to register, start and cancel
transactions. While it makes sense to expose the start RPC, it doesn't
really for the other two as they will only ever be used by Praefect,
which is also responsible for hosting the service. Furthermore, it's
hard to actually register and cancel transactions in Praefect if these
are exposed as services, only, as using them would require us to connect
to ourselves.

Revise the design of the transaction service by splitting it up into two
parts:

1. The transaction service, which now only exposes the "start" RPC to
   Gitaly nodes. The transaction handling logic has been split out of
   the service, bringing us to the second part.

2. The transaction manager. Similar to how e.g. the node manager works,
   this is where the actual business logic will take place. The
   transaction service gets a manager injected and will call out to it
   to serve the "start" RPC. The other two calls which were previously
   exposed via RPC will now be called directly on this manager by
   Praefect.

With this design, it becomes a lot easier to handle sessions on
Praefect's side.