transactions: Rework cancellation interface
When registering transactions, the assumption is that the registering
entity should also cancel the transaction at a later point in order to
clean up any data associated with it. This assumption is currently not
explicit in the API, and in fact we also delete the transaction when the
single node taking part in the transaction calls the `StartTransaction`
RPC. As a result, calling `CancelTransaction()` would result in an
error.

Improve the interface by removing `CancelTransaction()` and have
`RegisterTransaction()` return a cancellation function, making it
explicit that callers need to execute it.