git: Introduce new `CmdStream` structure
Passing standard input and output streams to commands is currently
fragile and hard to read, as we directly pass `io.Writer` and
`io.Reader` interfaces as arguments. This results in calls like
`unsafeBareCmd(ctx, nil, buf, nil, nil, "bare")`, which are really hard
to read as it's not clear which parameter does what.

Introduce a new `CmdStream` structure with fields for standard input,
output and error to increase readability.