Ignore order of remote branch names
For the `ls_remote_branches` experiment, we would log a warning if the
branch names in the control differed from the branch names in the
experiment.

However, if the arrays contained the same data (i.e., a successful
experiment) but in a different order, the two arrays were considered
different, but then when we calculated the difference, we ended up with
an empty array, and ended up logging an erroneous warning message.

Now the ultimate diff is the deciding factor of whether or not we log.
Two arrays of branch names with the same content but in different order
will not log a warning message, as we originally intended.