Added MidxRepack RPC
This patch laid out the basic machinery for creating a
multi-pack-index[1]. And use it to repack and cosolidate
multiple packs files without and data lost.

The basic flow of this process is as follow:
```
set git config

write
verify (and re-write on fail)

expire
verify (and re-write on fail)

repack
verify (and re-write on fail)
```

The expectation is that on the first run, `repack` will re-organize
deltas from multiple packfiles into 1 bigger packfile.
On the second run, `expire` will remove the old packfiles that has
already been repacked.

Note that due to safety reason discussed in [2], we dont want to call
`expire` immediately after repack. This has a storage size implication
as the new consolidated pack will contains all the delta from smaller
packfiles. But since all of these are delta compressed, and storage size
is generally cheap, it is negligible.

With `core.multiPackIndex=true`, git operations know to ultilize midx to
find which packfile to use. That means once repacked, the old packfiles
will not be used and are safe to removed.

MidxRepack respects most `git repack` config flags as of git 2.27.0,
with the exception of `repack.writeBitmaps`[3]. Therefore we should
NOT use this for a Full Repack for the time being.

Instead, we are using `--batch-size` option to make MidxRepack only
repack a subset of the smaller packfiles.
This should help reduce the number of packfiles inside a repository
without having to rewrite the bigggest packfile, thus its faster and
less disruptive than traditional `git repack -A -d`.

References:
[1]: https://git-scm.com/docs/git-multi-pack-index/
[2]: https://public-inbox.org/git/CAL3xRKfcKh=XxGso4DTRfJMAVMtwqyQkO0VUhqVuZUr_aQ5f+A@mail.gmail.com/T/#u
[3]: https://public-inbox.org/git/pull.626.git.1588684003766.gitgitgadget@gmail.com/