git: Default to latest supported version to build
Gitaly's Makefile can build Git, and currently defaults to master. For
local development this is inconvenient given there might be differences
which could lead to issues which are hard to debug.

The downside of this change, is that it locks the version, and requires
the Gitaly team to bump this version along with other Git version
upgrades.