Use bundler for version management for grpc_tools
The `grpc_tools` gem was installed using a `gem` invocation, which lead
to different versions of the same gem being used among developers. By
using the Gemfile for version management this should be resolved. The
gem is added as development dependency, and as such will not make its
way to production.

To support using bundler, `bundle exec` had to be used in the support
script, which required the execution context to be changed to the ruby
sub-directory. `Dir.chdir` is not thread safe, but currently I think
it's fair to assume this is not required.