gem: Upgrade github-linguist to 7.9.0
Going through the changelog for a different issue, I noted Gitaly was
lagging a few versions. Which includes new detected langauges as well as
a few performance improvements. Most notable, the shebang detection is
much quicker now as it doesn't iterate over the full file to just get
the first line.