hooks: Let pre-receive hook reach out to the transaction manager
In order to allow for strong consistency for Git reference updates,
Gitaly nodes are expected to perform voting on the updates they intend
to perform. While the server side for this exists already in form of the
transaction manager and reference transaction service, the client-side
is not yet wired up.

In this first iteration, we implement the client side by reaching out to
the reference transaction service from the pre-receive hook. While this
won't yet catch all reference updates but only those performed as the
result of git-push(1), it serves as a first minimally viable
demonstration that a two-phase commit via Git hooks works.

One interesting bit is that starting the transaction is _not_ guarded by
a feature flag on the hook side, but only by presence of environment
variables that carry transaction information. As this environment
variable is only set in case Praefect has registered a transaction,
which in turn only happens if the "reference_transactions" feature flag
is set, we are still guarded.

Further note the hook is only implemented in the Go hook service, which
is in turn guarded by the "hooks_rpc" feature flag. As the work is
currently of experimental nature, it's deemed as a nice proof of concept
for the Go hooks RPC.