Praefect: horizontal scaling of a single shard MVC
Random distribution of reads to up to date gitaly nodes.
Up to date status of node verified based on the state of the
replication queue. Backoff strategy is to use primary node
if there are no up to date secondaries or an error occurred.
This feature can be enabled with 'distribution_of_reads_enabled'.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2650