transactions: Log error cases
While we already log a transaction that is being started as well as a
successful transaction, we do not yet any error cases. Refactor the code
to put verification of transactions into its own function to unify all
current error cases into a single path and log any returned errors.