transaction: Add "node" field when logging committed transactions
The log message in case a transaction is committed is missing the "node"
field right now. In case we ever decide we want to enable transactions
in case a quorum was reached, then a subset of nodes will commit the
transaction while another subset will abort it. Thus, having the "node"
field logged in such a case may help quite a lot to see what's going on.