transactions: Implement metrics
We currently have no metrics in the transaction manager, making it hard
to estimate how many transaction succeed, how many fail and how long a
node has to wait until a quorum was reached and the transaction is
committed. This commit thus adds two metrics, one to count the number of
registered, started, invalid and committed transactions and one to
measure the delay between start and finish of the voting process.