hooks: Make transaction information available via environment
Now that we have transaction information passed through to us by
Praefect in case the reference transaction feature flag is turned on, we
need to pass it on to the Git hooks. Again, we follow the same approach
as chosen for the Praefect server connection information to make data
available via the environment of the Git hook.