makegen: Avoid hardcoding number of make jobs when building git
We currently hardcode the number of jobs to "8", which may be optimal on
some platforms, but definitely not on most. Some systems with a smaller
amount of CPU cores may thus be overloaded while others with a lot more
cores will be underloaded.

Let's improve this by using the number of cores available to the system,
plus one.