makegen: Fix building Git on musl-based platforms
By default, Git will assume that the regex implementation of libc
support the `REG_STARTEND` feature. As this feature started out as a
non-standard glibc extension, it's not available to all systems. Most
notably, musl libc doesn't implement it as it tries to be as strictly
POSIX compliant as it makes sense.

Fix the issue by amending NO_REGEX=YesPlease to Git's build options.
This will cause it to use its own embedded regex implementation which
obviously supports this feature.