Include Praefect usage in the usage ping
With introduction of praefect between GitLab and Gitaly it is
required to include info about Praefect into Usage Ping.
As Praefect is a transparent proxy between two listed above
the new parameter is added to the response and will be filled
by the Praefect after gathering info from Gitalies behind it.
Later it will be used to calculate the number of storages
to send statistics into Usage Ping.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2664