transactions: Extract transaction structure
Right now, transactions are still kind of simple in that they only allow
for single-node voting. As we're about to extend their scope to
multi-node voting, the associated code is going to grow more complex. As
a preparatory step, we thus extract the transaction-related part into
its own structure in order to nicely encapsulate it and not let the
transaction manager itself grow more complex.