danger: Remove warning to delete branches post-merge
A few minutes ago, I flipped the project setting that sets the 'Delete
source branch' tick by default when creating a new merge request. This
makes the danger check for the same thing irrelevant as the user chose
not to delete the branch if it's not set. Probably with good reason.