failover: Default to enabling SQL strategy
Given the SQL strategy is a reasonable default, and tested well enough,
this change makes the SQL strategy the preferred strategy. This was
`local`, which creates dataloss scenarios when multiple Praefects are
behind a load balancer.

Additionally, this changes behaviour around the read only feature. This
would default to being enabled. Currently availability is favoured over
consistency. Now this is flipped.

Fixes: https://gitlab.com/gitlab-org/gitaly/-/issues/2682