remember the previous write enabled primary node
Remembers the previous write enabled primary node. This will be used
to determine unreplicated writes from the previous write-enabled
primary after a failover.

A new column is needed as replication jobs do not always originate
from the primary node. Reconciling from a secondary to a read-only
primary in an attempt to get the complete data set to the primary
after a failover would produce replication jobs from the secondary
to the read-only primary. This makes it difficult to track the
previous write-enable primary solely from the replication jobs.