TreeEntries: reduce memory usage while parsing
While reading the source of GetTreeEntries, I found some issues with the
code as it used to read the full output buffer into memory, instead of
using an `io.Reader`. In this case the fix was to use an io.Reader, and
wrap that in a `bufio.NewReader()` to extend the API to it.