lines.Send() takes SenderOpts for grouping options
The lines.Send() allows for a command to be passed in, so the output can
be split based on a delimiter. To control what would be send to the
client this package will be extended to allow for pagination and limits.

The lines.Send() function than would read from the passed in io.Reader,
and write to the sink passed as `lines.Sender`, and the logic that
alters the execution between these would be send as part of the
SenderOpts struct.

A small test was added, mostly to force compilation of the package when
running `go test`, as well as unit test coverage.