Praefect: unhealthy nodes must not be used for read distribution
Coordinator must take into account state of the secondary node
that is considered for forwarding a read operation onto.
Unhealthy node must not be chosen even if it is in up to date
state in terms of replication.
Each node contains N last health checks that are used to verify
if node is healthy. It has been moved from `nodeCandidate`.
And `localElector` now does health check for it's nodes in
parallel.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2823