How to handle proxying FindRemoteRepository
It is a first steps towards making FindRemoteRepository RPC
to become STORAGE scoped. It is needed to simplify routing
in Praefect and omit need for Praefect to handle this RPC
as a special case. After change is done the request will be
routed into one of the Gitalies and handled there. The Git
is not needed to be installed on Prafect node in order to
support the implementation of only single RPC.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2442