Remove dependency on global var to detect repo location
Removes dependency on the global var config.Config to
detect location of the repository of storage at the
filesystem.
Locator interface includes set of methods from helper
package that are use in the services to get location of
the repository.
Implementation of the interface is injected into the
context via tags mechanism.
In order to break cyclic dependencies some of other funcs
was also moved.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699