Set a stable signature for .patch endpoints to create reproducible patches
This commit adds the `--signature` flag set to `GitLab` to the format-patch git
subcommand used in RawPatch of the Diff service.
This will result in the default signature of resulting .patch-files
to be replaced with `GitLab`, resulting in more stable patches, that will not change
file content based on the git version of the host (which is the default
signature appended to the patch). This is relevant to be able to
generate checksums or signatures for patches referenced via a GitLab
URL.

Implements #2842