Enforce literal paths for LastCommitForPath
By default, Git allows the pathspec to have glob patterns
(https://css-tricks.com/git-pathspecs-and-how-to-use-them), but the
LastCommitForPath RPC only uses the path argument as a literal. This can
be done via the --literal-pathspecs argument
(https://git-scm.com/docs/git#Documentation/git.txt---literal-pathspecs)
or by adding `(:literal)` before the pathspec. This commit does the
former because it seems clearer.

Closes https://gitlab.com/gitlab-org/gitaly/-/issues/2857